refactor: change type of telefone's field from smallint to int

+++ src/domain/migration/1643396225821-CreateManufacturerTable.ts
@@ -21,7 +21,7 @@ export class CreateManufacturerTable1643396225821 implements MigrationInterface 
                 },
                 {
                     name: 'telefone',
-                    type: 'smallint',
+                    type: 'int',
                     isNullable: true
                 },
                 {
